requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
pandas>=2.0.0
